aimbot/requirements.txt
opencv-python
numpy
keyboard
mss
pyserial